feat: allow setting a distance on the floating origin + camera fix
This allows the recalculation of the world to occur every x metres from origin - turns out recalculating the entire freaking world is rather expensive. Who knew?

Also fixed the camera to stop trying to do a bunch of mental math to stay in place and just ... update the local transform and move the script to the child camera object where it belongs. I AM SMART.