chore: tighten up layouts of bindings, begin pulling out expandable line
this is the start of making an expandable line prefab building block (not yet rolled into axis binding) - I'm thinking actually that some bindings will need this without being an axis and as such having an optional expand button in the layout would make sense. This is for things like flight assist hold vs toggle. Guess we'll see