feat: add state management for creating an offline server
OFFLINE: The server is started when the single player menu is invoked, setting max clients to 1. Whenever the main menu's top level menu is displayed, the server is shutdown if needed.

ONLINE: The server is started on host and not started for joining (previously it was?!), only the client is started.

This ensures there's always one server for all states.