optimisation: VFX particles no longer use Lit shader
this prevents a buildup of particles causing massive slowdown. It's a weird state to be in for sure but I can see this causing problems on lower end hardware, especially in multiplayer. Now uses fully unlit quads with some color tweaking which doesn't seem to impact FPS at all. Maybe that's naive.