refactor: make pause menu not utterly terrible
Some of the oldest code in the game right there, and you could tell.
This moves the cursor and event handling up a level to the whole screen space canvas and gives me a sane injection point to add an end-level screen. Additionally, it makes the pause menu use the menu base stuff for sanity's sake