feat: precache terrain in time trial around checkpoints for perf
this adds a shared tag between checkpoints and the local player - all terrain is generated around these markers, meaning the checkpoints act like players for the purposes of terrain generation. This has the trade-off that full terrain is rendered at distance but no stuttering during a race. There may be a better way to handle this (e.g. pinning tiles around checkpoints where main terrains are stored and swapped out from drafts as the camera nears) but this hopefully is a good stop-gap against stuttering.