fix: really really stupid painfully dumb fix to prevent geo reflecting
when the ship gets close to things the geo is exposed to the reflections from the ship. This is done with an axis-aligned bounding box so simply making the affected area does not fix this and there's no way to filter meshes from reflection probes in URP.

This fix adds six small, higher priority, skybox-only, one-time-render reflection probes around the ship which are applied to geometry instead of the realtime one.

I hate this so much